+++ lib/src/chart_interaction.dart
@@ -14,6 +14,7 @@ import 'values/value.dart';
 
 part 'chart_interaction_gesture_recognizers.dart';
 part 'chart_interaction_gesture_factory.dart';
+part 'chart_interaction_viewports.dart';
 
 /// [GChartInteractionHandler] for handling user interactions for the attached chart.
 // ignore: must_be_immutable
@@ -168,14 +169,9 @@ class GChartInteractionHandler {
     // hit test splitters
     for (int n = 0; n < _chart.panels.length; n++) {
       GPanel panel = _chart.panels[n];
-      if (n < _chart.panels.length - 1 &&
-          panel.splitterArea().contains(start)) {
-        GPanel? splitterPanel = _tryScalingSplitter(
-          n,
-          _chart.panels[n],
-          _chart.panels[n + 1],
-          start,
-        );
+      GPanel? nextPanel = _chart.nextVisiblePanel(startIndex: n + 1);
+      if (nextPanel != null && panel.splitterArea().contains(start)) {
+        GPanel? splitterPanel = _tryScalingSplitter(n, panel, nextPanel, start);
         if (splitterPanel != null) {
           _notify();
           return;
@@ -338,7 +334,8 @@ class GChartInteractionHandler {
       double h2 = panel2.panelArea().height;
       double moveToleranceMin = -panel1.graphArea().height + 50;
       double moveToleranceMax = panel2.graphArea().height - 50;
-      double weightUnit = panel1.heightWeight / (h1 / _chart.size.height);
+      double weightDensity =
+          (panel1.heightWeight + panel2.heightWeight) / (h1 + h2);
       _chart.splitter.resizingPanelIndex = panel1Index;
       _chart.crosshair.crossPosition.clear();
       _hookScaleUpdate = ({
@@ -352,9 +349,9 @@ class GChartInteractionHandler {
         );
         double h1New = h1 + moveDistance;
         double h2New = h2 - moveDistance;
-        panel1.heightWeight = h1New * weightUnit;
-        panel2.heightWeight = h2New * weightUnit;
-        _chart.layout(_chart.area);
+        panel1.heightWeight = h1New * weightDensity;
+        panel2.heightWeight = h2New * weightDensity;
+        _chart.resize(newArea: _chart.area, force: true);
       };
       _hookScaleEnd = (Velocity? velocity) {
         _chart.splitter.resizingPanelIndex = null;
@@ -581,191 +578,5 @@ class GChartInteractionHandler {
 
   void _notify() {
     _chart.repaint(layout: false);
-  }
-}
-
-/// Helper class for handling [GPointViewPort] interactions.
-class GPointViewPortInteractionHelper {
-  GPointViewPort? _pointViewPort;
-  final GRange _pointRangeScaling = GRange.empty();
-
-  bool get isScaling => _pointViewPort != null;
-
-  void interactionStart(GPointViewPort pointViewPort) {
-    _pointViewPort = pointViewPort;
-    _pointRangeScaling.copy(pointViewPort.range);
-  }
-
-  void interactionEnd() {
-    _pointRangeScaling.clear();
-    _pointViewPort?.selectedRange.clear();
-    _pointViewPort = null;
-    // _notify(finished: true);
-  }
-
-  void interactionZoomUpdate(Rect area, double zoomRatio) {
-    if (_pointViewPort == null || _pointRangeScaling.isEmpty) {
-      return;
-    }
-    _pointViewPort!.zoom(
-      area,
-      zoomRatio,
-      startRange: _pointRangeScaling,
-      centerPoint: _pointRangeScaling.last,
-      animate: false,
-      finished: false,
-    );
-  }
-
-  void interactionMoveUpdate(Rect area, double movedDistance) {
-    if (_pointViewPort == null || _pointRangeScaling.isEmpty) {
-      return;
-    }
-    double pointWidthStart = _pointViewPort!.pointSize(
-      area.width,
-      range: _pointRangeScaling,
-    );
-    double pointsMoved = movedDistance / pointWidthStart;
-    double startPointNew = _pointRangeScaling.begin! - pointsMoved;
-    double endPointNew = _pointRangeScaling.end! - pointsMoved;
-    _pointViewPort!.setRange(
-      startPoint: startPointNew,
-      endPoint: endPointNew,
-      finished: false,
-    );
-  }
-
-  void interactionSelectUpdate(
-    Rect area,
-    double startPosition,
-    double position, {
-    bool animation = true,
-    bool finished = false,
-  }) {
-    if (_pointViewPort == null || _pointRangeScaling.isEmpty) {
-      return;
-    }
-    final position1 = min(startPosition, position);
-    final position2 = max(startPosition, position);
-    _pointViewPort!.selectedRange.update(
-      _pointViewPort!.positionToPoint(
-        area,
-        max(area.left, min(area.right, position1)),
-      ),
-      _pointViewPort!.positionToPoint(
-        area,
-        max(area.left, min(area.right, position2)),
-      ),
-    );
-    if (finished) {
-      double value1 = _pointViewPort!.selectedRange.begin!;
-      double value2 = _pointViewPort!.selectedRange.end!;
-      double pointLeftNew = max(min(value1, value2), _pointRangeScaling.begin!);
-      double pointRightNew = min(max(value1, value2), _pointRangeScaling.end!);
-      double minPoints = area.width / _pointViewPort!.maxPointWidth;
-      if (pointRightNew - pointLeftNew < minPoints) {
-        // adjust to make sure the width is not less than minPointWidth
-        double center = (pointLeftNew + pointRightNew) / 2;
-        pointLeftNew = center - minPoints / 2;
-        pointRightNew = center + minPoints / 2;
-      }
-      _pointViewPort!.animateToRange(
-        GRange.range(pointLeftNew, pointRightNew),
-        finished,
-        animation,
-      );
-    }
-  }
-}
-
-/// Helper class for handling [GValueViewPort] interactions.
-class GValueViewPortInteractionHelper {
-  GValueViewPort? _valueViewPort;
-  final GRange _rangeScaling = GRange.empty();
-
-  bool get isScaling => _valueViewPort != null;
-
-  void interactionStart(GValueViewPort valueViewPort) {
-    _valueViewPort = valueViewPort;
-    _rangeScaling.copy(valueViewPort.range);
-  }
-
-  void interactionEnd() {
-    _rangeScaling.clear();
-    _valueViewPort?.selectedRange.clear();
-    _valueViewPort = null;
-    // _notify(finished: true);
-  }
-
-  void interactionZoomUpdate(
-    Rect area,
-    double zoomRatio, {
-    bool notify = true,
-  }) {
-    if (!isScaling) {
-      return;
-    }
-    _valueViewPort!.zoom(
-      area,
-      zoomRatio,
-      startRange: _rangeScaling,
-      animate: false,
-      finished: true,
-      notify: notify,
-    );
-  }
-
-  void interactionMoveUpdate(Rect area, double movedDistance) {
-    if (!isScaling) {
-      return;
-    }
-    double valueMoved =
-        (movedDistance / area.height) *
-        (_rangeScaling.last! - _rangeScaling.first!);
-    double endValueNew = _rangeScaling.last! + valueMoved;
-    double startValueNew = _rangeScaling.first! + valueMoved;
-    _valueViewPort!.setRange(startValue: startValueNew, endValue: endValueNew);
-  }
-
-  void interactionSelectUpdate(
-    Rect area,
-    double startPosition,
-    double position, {
-    bool finished = false,
-  }) {
-    if (!isScaling) {
-      return;
-    }
-    final value1 = _valueViewPort!.positionToValue(
-      area,
-      max(area.top, min(area.bottom, startPosition)),
-    );
-    final value2 = _valueViewPort!.positionToValue(
-      area,
-      max(area.top, min(area.bottom, position)),
-    );
-    _valueViewPort!.selectedRange.update(
-      min(value1, value2),
-      max(value1, value2),
-    );
-    if (finished) {
-      double endValueNew = min(
-        _valueViewPort!.selectedRange.last!,
-        _rangeScaling.last!,
-      );
-      double startValueNew = max(
-        _valueViewPort!.selectedRange.first!,
-        _rangeScaling.first!,
-      );
-      (startValueNew, endValueNew) = _valueViewPort!.clampScaleRange(
-        startValueNew,
-        endValueNew,
-      );
-      _valueViewPort!.animateToRange(
-        GRange.range(startValueNew, endValueNew),
-        finished,
-        true,
-      );
-    }
   }
 }